TrainingRegistar entity moved to domain.layer
